fix:chatgpt-model
